1075: implement Design
