+++ administration/package.json
@@ -7,7 +7,7 @@
   },
   "dependencies": {
     "@apollo/client": "^3.7.16",
-    "@blueprintjs/core": "^5.0.0",
+    "@blueprintjs/core": "^5.3.0",
     "@blueprintjs/select": "^5.0.0",
     "@blueprintjs/table": "^5.0.0",
     "@bufbuild/protobuf": "^1.2.1",
