+++ administration/src/project-configs/bayern/config.ts
@@ -5,10 +5,10 @@ import { ProjectConfig } from '../getProjectConfig'
 import { DataPrivacyAdditionalBaseText, DataPrivacyBaseText, dataPrivacyBaseHeadline } from './dataPrivacyBase'
 import pdfConfiguration from './pdf'
 
-export const applicationJsonToCardQuery = (json: JsonField<'Array'>): string | null => {
-  const query = new URLSearchParams()
+export const applicationJsonToPersonalData = (
+  json: JsonField<'Array'>
+): { forenames?: string; surname?: string } | null => {
   const personalData = findValue(json, 'personalData', 'Array')
-  const cardType = findValue(json, 'cardType', 'String')
 
   if (!personalData) {
     return null
@@ -17,11 +17,21 @@ export const applicationJsonToCardQuery = (json: JsonField<'Array'>): string | n
   const forenames = findValue(personalData, 'forenames', 'String')
   const surname = findValue(personalData, 'surname', 'String')
 
-  if (!forenames || !surname || !cardType) {
+  return { forenames: forenames?.value, surname: surname?.value }
+}
+
+export const applicationJsonToCardQuery = (json: JsonField<'Array'>): string | null => {
+  const query = new URLSearchParams()
+  const applicationDetails = findValue(json, 'applicationDetails', 'Array') ?? json
+  const cardType = findValue(applicationDetails, 'cardType', 'String')
+
+  const personalData = applicationJsonToPersonalData(json)
+
+  if (!personalData || !personalData.forenames || !personalData.surname || !cardType) {
     return null
   }
 
-  query.set(config.card.nameColumnName, `${forenames.value} ${surname.value}`)
+  query.set(config.card.nameColumnName, `${personalData.forenames} ${personalData.surname}`)
   const cardTypeExtensionIdx = config.card.extensions.findIndex(ext => ext === BavariaCardTypeExtension)
   const value = cardType.value === 'Goldene Ehrenamtskarte' ? 'Goldkarte' : 'Standard'
   query.set(config.card.extensionColumnNames[cardTypeExtensionIdx] ?? '', value)
@@ -33,6 +43,7 @@ const config: ProjectConfig = {
   name: 'Ehrenamtskarte Bayern',
   projectId: 'bayern.ehrenamtskarte.app',
   applicationFeature: {
+    applicationJsonToPersonalData,
     applicationJsonToCardQuery,
   },
   staticQrCodesEnabled: false,
