+++ administration/src/project-configs/showcase/config.ts
@@ -1,6 +1,6 @@
 import BavariaCardTypeExtension from '../../cards/extensions/BavariaCardTypeExtension'
 import RegionExtension from '../../cards/extensions/RegionExtension'
-import bayern, { applicationJsonToCardQuery } from '../bayern/config'
+import bayern, { applicationJsonToCardQuery, applicationJsonToPersonalData } from '../bayern/config'
 import { DataPrivacyBaseText, dataPrivacyBaseHeadline } from '../bayern/dataPrivacyBase'
 import { ProjectConfig } from '../getProjectConfig'
 
@@ -8,6 +8,7 @@ const config: ProjectConfig = {
   name: 'Showcase Berechtigungskarte',
   projectId: 'showcase.entitlementcard.app',
   applicationFeature: {
+    applicationJsonToPersonalData,
     applicationJsonToCardQuery,
   },
   staticQrCodesEnabled: false,
